refactor:페이지에 대한 component 파일 병합
