+++ moem-app/src/App.jsx
@@ -1,5 +1,3 @@
-import logo from './assets/logo.png'
-import HomeButton from './components/HomeButton'
 import LoginPage from "./pages/Loginpage";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import './App.css'
@@ -7,18 +5,18 @@ import Layout from './components/layouts/layout';
 import MatchingPage from './pages/MatchingPage';
 import RecruitmentPage from './pages/RecruitmentPage';
 import MatchingDetailPage from './pages/MatchingDetailPage';
-import { Signuppage } from './pages/Signuppage';
-import { Mainpage } from './pages/Mainpage';
-import Home from "./pages/Home";
+import LandingPage from "./pages/LandingPage";
+import DashBoard from './pages/DashBoard';
+import { SignupPage } from './pages/SignupPage';
 
 function App() {
   return (
     <Layout>
       <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<LandingPage/>} />
         <Route path="/login" element={<LoginPage />} />
-        <Route path='/signup' element={<Signuppage/>}/>
-        <Route path='/main' element={<Mainpage/>}/>
+        <Route path='/signup' element={<SignupPage/>}/>
+        <Route path='/main' element={<DashBoard/>}/>
         <Route path='/recruitments' element={<MatchingPage/>} />
         <Route path='/recruitments/:id' element={<MatchingDetailPage/>} />
         <Route path='/recruitments/new' element={<RecruitmentPage/>} />
