+++ moem-app/src/pages/DashBoard.jsx
@@ -1,14 +1,18 @@
 import React from 'react'
 import logo from '../assets/logo.png'
-import Maincomponent from '../components/maincomponent'
+import { Link } from 'react-router-dom';
 
 export const DashBoard = () => {
   return (
     <div className='header-container'>
         <img src={logo} className="logo-img"/>
-        <div >
-        <Maincomponent/>
-        </div>
+         <div class='btn'>
+//       <nav>
+//         <Link to ="/login">매칭</Link>
+//         <Link to ="/">로그아웃</Link>
+//         <Link to ="/login">마이페이지</Link>
+//         </nav>
+//     </div>
     </div>
   )
 }
